minor NY fix

+++ asm_src/libft/ft_atoi.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: swedde <swedde@student.42.fr>              +#+  +:+       +#+        */
+/*   By: nsheev <nsheev@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/14 18:15:49 by nsheev            #+#    #+#             */
-/*   Updated: 2019/12/16 23:26:28 by swedde           ###   ########.fr       */
+/*   Updated: 2020/01/14 15:55:06 by nsheev           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	unsigned long long int	i;
 	int						neg;
